% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsi.R
\docType{data}
\name{HSI}
\alias{HSI}
\title{HSI time series data.}
\format{
A data frame with Date, Price variables:
\describe{
  \item{Date}{Trading day}
  \item{Close}{Closing price}
}
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
HSI
}
\description{
A dataset of 1122 rows containing the prices and Date attributes of
HANG SENG INDEX (^HSI) from January 2, 2004 to June 30, 2008.
}
\keyword{datasets}
