% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppl-package.R
\docType{package}
\name{lppl-package}
\alias{lppl}
\alias{lppl-package}
\title{lppl: Log Periodic Power Low(LPPL)}
\description{
The package implements Log-Periodic-Power-Low(LPPL) models proposed to test for financial bubbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcgyver3/lppl}
  \item Report bugs at \url{https://github.com/mcgyver3/lppl/issues}
}

}
\author{
\strong{Maintainer}: Hyun Nam \email{mcgyver3@gmail.com} (\href{https://orcid.org/0000-0002-8464-0173}{ORCID})

}
\keyword{LPPL}
