% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_crashtime.R
\name{lppl.find_crashtime}
\alias{lppl.find_crashtime}
\title{lppl.find_crashtime}
\usage{
lppl.find_crashtime(
  tradedate,
  price,
  Left_days = 262,
  Right_days = 60,
  Right_scale = 0.8,
  showplot = TRUE
)
}
\arguments{
\item{tradedate}{Use a "Date" type vector as the transaction date.}

\item{price}{Closing price numeric data vector}

\item{Left_days}{The peak of critical time is higher than any peak prior to 262 days (Default : 262)}

\item{Right_days}{It implements that the peak should drop by 25\% during the right_days (60 days) period, and the default value is 60.}

\item{Right_scale}{set the percentage of drop price from peak price. (Default : 0.8)}
}
\value{

}
\description{
This is the summation to find the critical point in the entered world thermal data.
}
\details{
There are three conditions at the critical point.
During the 262 days prior to the first critical point, the highest price must be less than the current price.
Prices must drop by 25\% or more within 60 days prior to the second critical point.
Third, there should be no price higher than the serial price for 60 days after the critical point.
}
\examples{

library(lppl)

tradedate <- lppl::NASDAQ$Date
price <- lppl::NASDAQ$Close

dfCRASH <- lppl.find_crashtime(tradedate =tradedate, price = price)

}
\references{
D.S. Brée and N. L. Joseph, “Testing for financial crashes using the log-periodic power law model,” International Review of Financial Analysis, vol. 30, pp. 287–297, 2013.
}
\keyword{crashtime}
\keyword{lppl,}
