% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasdaq.R
\docType{data}
\name{NASDAQ}
\alias{NASDAQ}
\title{NASDAQ time series data.}
\format{
A data frame with Date, Price variables:
\describe{
  \item{Date}{Trading day}
  \item{Close}{Closing price}
}
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
NASDAQ
}
\description{
A dataset of 1,009 rows containing the prices and Date attributes of
NASDAQ from July 1, 1997 to June 29, 2001.
}
\keyword{datasets}
