% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nikkei.R
\docType{data}
\name{NIKKEI}
\alias{NIKKEI}
\title{NIKKEI time series data.}
\format{
A data frame with Date, Price variables:
\describe{
  \item{Date}{Trading day}
  \item{Close}{Closing price}
}
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
NIKKEI
}
\description{
A dataset of 1043 rows containing the prices and Date attributes of
NIKKEI INDEX from July 1, 1987 to June 28, 1991.
}
\keyword{datasets}
