% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssec.R
\docType{data}
\name{SSEC}
\alias{SSEC}
\title{SSEC time series data.}
\format{
A data frame with Date, Price variables:
\describe{
  \item{Date}{Trading day}
  \item{Close}{Closing price}
}
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
SSEC
}
\description{
A dataset of 968 rows containing the prices and Date attributes of
Shanghai Composite Index from January 4, 2012 to December 29, 2015.
}
\keyword{datasets}
